refactor(qa): relocate Mutation schema out of corpus tree (arc PR3)
